+++ CHANGELOG.md
@@ -2,13 +2,39 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-### [0.2.9-ga.1](https://github.com/salesforcecli/plugin-functions/compare/v0.2.9-ga.0...v0.2.9-ga.1) (2021-06-28)
+### [0.2.14](https://github.com/salesforcecli/plugin-functions/compare/v0.2.12...v0.2.14) (2021-06-29)
 
-### Features
+### Bug Fixes
 
-- add findDeployers hook ([#101](https://github.com/salesforcecli/plugin-functions/issues/101)) ([e434a5e](https://github.com/salesforcecli/plugin-functions/commit/e434a5e07ff2bcd9811b347ae855fbc5eca55b21))
+- add --extended hidden flag for debugging ([#96](https://github.com/salesforcecli/plugin-functions/issues/96)) ([ff187b8](https://github.com/salesforcecli/plugin-functions/commit/ff187b8bbf82c4acd849878152b0e2ba12258426))
+- batch reference deletion in env:delete ([#98](https://github.com/salesforcecli/plugin-functions/issues/98)) ([7f4fb6a](https://github.com/salesforcecli/plugin-functions/commit/7f4fb6a2195ee34e93ddbe327ddd80dd8ea3097f))
 
-### [0.2.9-ga.0](https://github.com/salesforcecli/plugin-functions/compare/v0.2.8...v0.2.9-ga.0) (2021-06-22)
+### [0.2.13](https://github.com/salesforcecli/plugin-functions/compare/v0.2.12...v0.2.13) (2021-06-24)
+
+### Bug Fixes
+
+- add --extended hidden flag for debugging ([#96](https://github.com/salesforcecli/plugin-functions/issues/96)) ([ff187b8](https://github.com/salesforcecli/plugin-functions/commit/ff187b8bbf82c4acd849878152b0e2ba12258426))
+- batch reference deletion in env:delete ([#98](https://github.com/salesforcecli/plugin-functions/issues/98)) ([7f4fb6a](https://github.com/salesforcecli/plugin-functions/commit/7f4fb6a2195ee34e93ddbe327ddd80dd8ea3097f))
+
+### [0.2.12](https://github.com/salesforcecli/plugin-functions/compare/v0.2.11...v0.2.12) (2021-06-24)
+
+### Bug Fixes
+
+- In run function, stop the cli action on a successful request ([#97](https://github.com/salesforcecli/plugin-functions/issues/97)) ([463d111](https://github.com/salesforcecli/plugin-functions/commit/463d111cf1f5e5978bf78b9774a1801d204f256b))
+
+### [0.2.11](https://github.com/salesforcecli/plugin-functions/compare/v0.2.10...v0.2.11) (2021-06-24)
+
+### [0.2.10](https://github.com/salesforcecli/plugin-functions/compare/v0.2.9...v0.2.10) (2021-06-24)
+
+### Bug Fixes
+
+- **cli:** use function references for env:display function list rather than local filesystem ([#93](https://github.com/salesforcecli/plugin-functions/issues/93)) ([9558779](https://github.com/salesforcecli/plugin-functions/commit/9558779a02c653c5e00d486bed9e0f0b5fdb17ef))
+
+### [0.2.9](https://github.com/salesforcecli/plugin-functions/compare/v0.2.8...v0.2.9) (2021-06-24)
+
+### Bug Fixes
+
+- upsert function references sequentially ([#94](https://github.com/salesforcecli/plugin-functions/issues/94)) ([1970264](https://github.com/salesforcecli/plugin-functions/commit/1970264a6be0514ed5cb93a02c9ff10cffc39dc9))
 
 ### [0.2.8](https://github.com/salesforcecli/plugin-functions/compare/v0.2.7...v0.2.8) (2021-06-22)